requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bce-python-sdk==0.9.5
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
decorator==5.1.1
deepdiff==6.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google-auth==2.28.1
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==6.11.0
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.22
lazy_loader==0.3
llama-index==0.10.16
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.7
llama-index-core==0.10.16.post1
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.8
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.5
llama-parse==0.3.6
llamaindex-py-client==0.1.13
lmdb==1.4.1
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
paddlepaddle==2.6.0
pandas==2.2.1
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==10.2.0
posthog==3.5.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.8
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
PyMuPDF==1.23.21
pyparsing==3.1.1
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-iso639==2024.2.7
python-magic==0.4.27
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
rarfile==4.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.12.5
unstructured-client==0.21.1
urllib3==2.2.1
uvicorn==0.27.1
visualdl==2.5.3
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
llama-index-indices-managed-vectara==0.1.4
PyPDF2==3.0.1